// src/generated/models/Sfsures_usergroupassignmentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Sfsures_usergroupassignmentsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Sfsures_usergroupassignmentsstatecode = keyof typeof Sfsures_usergroupassignmentsstatecode;
export const Sfsures_usergroupassignmentsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Sfsures_usergroupassignmentsstatuscode = keyof typeof Sfsures_usergroupassignmentsstatuscode;

export interface Sfsures_usergroupassignmentsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  "sfsures_Group@odata.bind"?: string;
  sfsures_name: string;
  "sfsures_User@odata.bind"?: string;
  sfsures_usergroupassignmentid: string;
  statecode: Sfsures_usergroupassignmentsstatecode;
  statuscode?: Sfsures_usergroupassignmentsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Sfsures_usergroupassignments extends Sfsures_usergroupassignmentsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  sfsures_groupname?: string;
  sfsures_username?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
  sfsures_group?: object;
  _sfsures_group_value?: string;
  sfsures_user?: object;
  _sfsures_user_value?: string;
}